Fix NextAuth configuration and build errors
Resolved Next.js App Router compatibility issues:

• Moved authOptions from route file to dedicated lib/auth.ts
• Updated NextAuth route to only export HTTP handlers
• Added Suspense boundary for useSearchParams in auth error page
• Fixed build failures related to invalid route exports

This ensures proper separation of concerns and Next.js 15 compatibility.
